+++ BasicMemoryInterface.h
@@ -13,9 +13,9 @@ public:
     ~BasicMemoryInterface();
 
     bool UpdateProcessId(const wchar_t* processName);
-    bool ReadRaw(uint64_t address, void* pBuffer, unsigned long size);
-    bool WriteRaw(uint64_t address, void* pBuffer, unsigned long size);
-    uint64_t GetModuleBase();
+    bool ReadRaw(uintptr_t address, void* pBuffer, unsigned long size);
+    bool WriteRaw(uintptr_t address, void* pBuffer, unsigned long size);
+    uintptr_t GetModuleBase();
 private:
     void* handle;
     unsigned long pid;
