+++ GameWorld.cpp
@@ -6,12 +6,12 @@
 #include "mdissect/mdissect.hpp"
 
 GameWorld GameWorld::Get() {
-	uint64_t gameWorldAddressGameObject = Global::gom.GetActiveObjects().GetObject("GameWorld");
+	uintptr_t gameWorldAddressGameObject = Global::gom.GetActiveObjects().GetObject("GameWorld");
 	if (!gameWorldAddressGameObject) {
 		return GameWorld{0};
 	}
 
-	uint64_t gameWorldAddress = Memory::ReadChain<uint64_t>(
+	uintptr_t gameWorldAddress = Memory::ReadChain<uintptr_t>(
 		Global::pMemoryInterface,
 		gameWorldAddressGameObject,
 		{0x30, 0x18, 0x28}
@@ -23,12 +23,12 @@ GameWorld GameWorld::Get() {
 std::vector<Player>& GameWorld::GetPlayers() {
 	if (!this->cachedPlayers.has_value()) {
 		std::vector<Player> players{};
-		uint64_t registeredPlayersAddress = Memory::ReadValue<uint64_t>(
+		uintptr_t registeredPlayersAddress = Memory::ReadValue<uintptr_t>(
 			Global::pMemoryInterface,
 			this->address + Offsets::GameWorld::RegisteredPlayers
 		);
 		if (registeredPlayersAddress) {
-			std::vector<uint64_t> playerAddresses = Memory::ReadList<uint64_t>(Global::pMemoryInterface, registeredPlayersAddress);
+			std::vector<uintptr_t> playerAddresses = Memory::ReadList<uintptr_t>(Global::pMemoryInterface, registeredPlayersAddress);
 			for (int i = 0; i < playerAddresses.size(); i++) {
 				players.push_back(Player{ playerAddresses[i]});
 			}
@@ -43,7 +43,7 @@ std::vector<Player>& GameWorld::GetPlayers() {
 std::vector<WorldLootItem>& GameWorld::GetLoot() {
 	size_t filtersCount = Global::pSettings->lootESP.filters.size();
 	if (!this->cachedLoot.has_value()) {
-		std::vector<uint64_t> addresses = Memory::ReadList<uint64_t>(Global::pMemoryInterface, Memory::ReadValue<uint64_t>(Global::pMemoryInterface, this->address + 0x80));
+		std::vector<uintptr_t> addresses = Memory::ReadList<uintptr_t>(Global::pMemoryInterface, Memory::ReadValue<uintptr_t>(Global::pMemoryInterface, this->address + 0x80));
 		size_t length = addresses.size();
 		std::vector<WorldLootItem> loot{};
 		loot.reserve(length);
